// week-4/enterprise-composer-app/src/app/composer.interface.ts
// Title: Exercise 3.2 - Passing Data to Routes pt. 1
// Author: Professor Krasso
// Date: 6 Nov 2022
// Modified By: Kayla McDanel
// Description: Enhanced Composer App
// Code Attribution: Code and instruction provided by Professor Krasso's videos and assignment docs.


export interface IComposer {
  composerId: number;
  fullName: string;
  genre: string;
}